LayoutTests/fast/frames/seamless/seamless-sandbox-flag-expected.txt

PASS iframe.seamless is true
PASS nestedFrame.seamless is true
FAIL window.getComputedStyle(iframe).width should be 200px. Was -4px.
FAIL window.getComputedStyle(iframe).height should be 150px. Was 0px.
